+++ CHANGELOG.md
@@ -25,6 +25,9 @@
     - `{episode_title_clean}`- Clean title parsed from media databases.
     - `{episode_title_exact}`- Title parsed from media databases with no modifications.
     - `{episode_title_exact}`- Title parsed from media databases with no modifications.
+    - `{episode_mediainfo}` - Synopsis of all episodes mediainfo.
+    - `{episode_metadata}` - Synopsis of all episodes metadata.
+    - `{episode_metadata_mediainfo}` - Synopsis of all episodes metadata + mediainfo.
   - **NfoTokens**:
     - `{total_seasons}`- Total seasons in series.
     - `{total_episodes}`- Total episodes in season/series.